Fix doc examples by moving to separate .py file
