+++ mkdocs.yml
@@ -34,6 +34,8 @@ markdown_extensions:
   - admonition
   - pymdownx.details
   - pymdownx.superfences
+  - pymdownx.snippets:
+      dedent_subsections: true
   - markdown_include.include
   - toc:
       permalink: "#"
@@ -63,5 +65,6 @@ plugins:
 
 watch:
   - CHANGELOG.md
+  - docs_src
   - README.md
   - tanchan
