+++ pyproject.toml
@@ -99,6 +99,7 @@ ignore = [
     "E231",    # missing whitespace after ','
 ]
 per-file-ignores = [
+    "docs_src/*.py: A001, D100, D103, D407, FA101, F841, INP001, TC002, TYP006",
     "noxfile.py: D100, FA101, F401, F403, INP001",
     "tanchan/py.typed: D100",
     "tests/*.py: CCE002, D100, D101, D103, D104, FA100, FA101",
@@ -147,7 +148,7 @@ mypy_allowed_to_fail = true
 mypy_targets = ["./tanchan"]
 path_ignore = "tanchan\\/_internal\\/vendor\\/"
 project_name = "tanchan"
-top_level_targets = ["./noxfile.py", "./tanchan", "./tests"]
+top_level_targets = ["./docs_src", "./noxfile.py", "./tanchan", "./tests"]
 
 [tool.piped.github_actions.freeze_for_pr]
 [tool.piped.github_actions.lint]
@@ -168,7 +169,7 @@ python_versions = ["3.9", "3.10", "3.11"]
 
 [tool.pyright]
 exclude = ["tanchan/_internal/vendor"]
-include = ["tanchan", "noxfile.py", "tests"]
+include = ["docs_src", "tanchan", "noxfile.py", "tests"]
 
 pythonVersion = "3.9"
 typeCheckingMode = "strict"
